Move RageModel to RageUtil/Graphics/Model
