+++ src/RageUtil/CMakeLists.txt
@@ -53,7 +53,7 @@ list(APPEND SMDATA_RAGE_GRAPHICS_SRC
   "Graphics/Display/OpenGL/FrameBufferRenderTargetGL.cpp"
   "Graphics/Display/OpenGL/ShaderGL.cpp"
   "Graphics/Display/OpenGL/GLHelper.cpp"
-  "Graphics/RageModelGeometry.cpp"
+  "Graphics/Display/Model/RageModelGeometry.cpp"
   "Graphics/Display/Surface/RageSurface.cpp"
   "Graphics/Display/Surface/RageSurface_Load.cpp"
   "Graphics/Display/Surface/RageSurface_Load_XPM.cpp"
@@ -78,7 +78,7 @@ list(APPEND SMDATA_RAGE_GRAPHICS_HPP
   "Graphics/Display/OpenGL/FrameBufferRenderTargetGL.hpp"
   "Graphics/Display/OpenGL/ShaderGL.hpp"
   "Graphics/Display/OpenGL/GLHelper.hpp"
-  "Graphics/RageModelGeometry.h"
+  "Graphics/Display/Model/RageModelGeometry.h"
   "Graphics/Display/Surface/RageSurface.h"
   "Graphics/Display/Surface/RageSurface_Load.h"
   "Graphics/Display/Surface/RageSurface_Load_XPM.h"
