+++ src/RageUtil/Graphics/Display/Model/RageModelGeometry.cpp
