+++ src/RageUtil/Graphics/Display/Model/RageModelGeometry.h
